tools/arr-sync/requirements.txt
requests>=2.28.0
pyyaml>=6.0
click>=8.0
